requirements.txt
flask==1.1.2
SQLAlchemy-1.3.20 
flask-sqlalchemy-2.4.4